+++ src/app/components/header/header.component.ts
@@ -1,7 +1,7 @@
 import { Component, WritableSignal } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { RouterLink, RouterLinkActive } from '@angular/router';
-import { UserService } from '../core/services/user/user.service';
+import { UserService } from '../../core/services/user/user.service';
 
 @Component({
   selector: 'app-header',
