+++ src/app/components/pages/carbon-footprint/carbon-footprint.component.scss
