+++ src/app/components/pages/carbon-footprint/carbon-footprint.component.ts
@@ -2,8 +2,8 @@ import { Component, signal, Signal } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MatButtonModule } from '@angular/material/button';
 import { MatTableModule } from '@angular/material/table';
-import { Voyage } from '../../core/services/carbon-footprint-compute/carbon-footprint-compute.types';
-import { CarbonFootprintComputeService } from '../../core/services/carbon-footprint-compute/carbon-footprint-compute.service';
+import { Voyage } from '../../../core/services/carbon-footprint-compute/carbon-footprint-compute.types';
+import { CarbonFootprintComputeService } from '../../../core/services/carbon-footprint-compute/carbon-footprint-compute.service';
 import { CarbonFootprintFormComponent } from './components/carbon-footprint-form/carbon-footprint-form.component';
 import { MatChipsModule } from '@angular/material/chips';
 import { NgxSkeletonLoaderModule } from 'ngx-skeleton-loader';
