+++ src/app/components/pages/carbon-footprint/components/carbon-footprint-result/carbon-footprint-result.component.ts
