+++ src/app/components/pages/profile/profile.component.ts
@@ -1,5 +1,5 @@
 import { Component, WritableSignal } from '@angular/core';
-import { UserService } from '../../core/services/user/user.service';
+import { UserService } from '../../../core/services/user/user.service';
 
 @Component({
   selector: 'app-profile',
