+++ src/app/core/guards/login.guard.ts
@@ -1,6 +1,6 @@
 import { inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
-import { UserService } from './core/services/user/user.service';
+import { UserService } from '../services/user/user.service';
 
 export const loginGuard: CanActivateFn = (_route, _state) => {
   const userService = inject(UserService);
